Consolidate ICM architecture: eliminate redundancy and defensive imports
- Consolidated duplicate ICM implementations into single clean module
- Removed all defensive NCLONE_AVAILABLE checks, assume nclone always available
- Renamed nclone_integration.py to reachability_exploration.py for clarity
- Standardized naming: ICMNetwork, ICMTrainer (removed 'Consolidated' prefixes)
- Added backward compatibility aliases for existing code
- Consolidated test files into single test_reachability_icm.py
- Created comprehensive ICM_INTEGRATION_GUIDE.md documentation
- Updated all imports and references to use clean architecture
- Validated: All tests pass, performance requirements met

Architecture improvements:
- Reduced code from 728 to 381 lines in main ICM module
- Eliminated redundant modules and defensive patterns
- Clean integration with nclone reachability systems
- Maintains <1ms computation performance requirements
