+++ npp_rl/intrinsic/icm.py
@@ -1,728 +1,386 @@
 """
-Intrinsic Curiosity Module (ICM) with Reachability Integration
+Reachability-Aware Intrinsic Curiosity Module (ICM)
 
-This module enhances the existing Intrinsic Curiosity Module (ICM) with reachability-aware
-exploration that avoids wasting curiosity on unreachable areas, improving sample efficiency.
+This module provides a clean implementation that integrates with nclone's 
+existing reachability systems for enhanced exploration in RL training.
 
-Integration Strategy:
-- Reachability scaling modulates base curiosity based on accessibility analysis
-- Frontier detection boosts exploration of newly reachable areas from nclone physics
-- Strategic weighting prioritizes exploration near level completion objectives
-- Performance-optimized for real-time RL training (<1ms curiosity computation)
+Key Design Principles:
+1. Use nclone's ExplorationRewardCalculator as the base exploration system
+2. Enhance it with reachability awareness rather than replacing it
+3. Leverage nclone's OpenCV-based reachability analysis and frontier detection
+4. Avoid duplicating exploration tracking logic
+5. Maintain performance requirements (<1ms computation)
 
-Theoretical Foundation:
-- Base ICM: Pathak et al. (2017) "Curiosity-driven Exploration by Self-supervised Prediction"  
-- Reachability guidance: Ecoffet et al. (2019) "Go-Explore: a New Approach for Hard-Exploration Problems"
-- Strategic exploration: Burda et al. (2018) "Exploration by Random Network Distillation"
-- Frontier exploration: Stanton & Clune (2018) "Deep curiosity search"
-
-nclone Integration:
-- Uses compact reachability features from nclone.graph.reachability for accessibility assessment
-- Integrates with TieredReachabilitySystem for performance-optimized reachability queries
-- Maintains compatibility with existing NPP physics constants and level objectives
+Integration with nclone:
+- TieredReachabilitySystem for multi-tier reachability analysis
+- CompactReachabilityFeatures for 64-dimensional feature encoding
+- FrontierDetector for boundary detection and classification
+- ExplorationRewardCalculator for multi-scale exploration tracking
 """
 
-# Standard library imports
-import math
-import time
-from collections import deque
-from typing import Dict, Tuple, Optional, Set, List, Any
-
-# Third-party imports
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import numpy as np
+import time
+from typing import Dict, Any, Optional, Tuple, List
+from collections import deque
 
-# Local imports
-from .nclone_integration import ReachabilityAwareExplorationCalculator, NCLONE_AVAILABLE
-
-
-class ExplorationHistory:
-    """
-    Tracks exploration history for frontier detection and strategic planning.
-    
-    This component maintains a memory of visited positions and reachable areas
-    to support frontier detection and strategic exploration weighting.
-    
-    References:
-    - Memory-based exploration: Ecoffet et al. (2019) "Go-Explore"
-    - Frontier detection: Stanton & Clune (2018) "Deep curiosity search"
-    """
-    
-    def __init__(self, max_size: int = 10000):
-        """Initialize exploration history with bounded memory."""
-        self.max_size = max_size
-        self.visited_positions = deque(maxlen=max_size)
-        self.reachable_areas = deque(maxlen=max_size)
-        self.timestamps = deque(maxlen=max_size)
-        
-    def add_position(self, position: Tuple[float, float], reachable_area: Set[Tuple[int, int]], timestamp: float):
-        """Add a visited position with its reachable area."""
-        self.visited_positions.append(position)
-        self.reachable_areas.append(reachable_area)
-        self.timestamps.append(timestamp)
-    
-    def get_recent_positions(self, window: int = 1000) -> List[Tuple[float, float]]:
-        """Get recent visited positions."""
-        return list(self.visited_positions)[-window:]
-    
-    def clear(self):
-        """Clear exploration history."""
-        self.visited_positions.clear()
-        self.reachable_areas.clear()
-        self.timestamps.clear()
-
-
-class FrontierDetector:
-    """
-    Detects exploration frontiers - newly accessible areas after state changes.
-    
-    This component identifies areas that become reachable after switch activations
-    or other state changes, enabling targeted exploration of newly accessible regions.
-    
-    Based on frontier-based exploration from Stanton & Clune (2018) "Deep curiosity search"
-    and the accessibility principles from Ecoffet et al. (2019) "Go-Explore".
-    """
-    
-    def __init__(self, reachability_dim: int = 64, memory_size: int = 1000):
-        """Initialize frontier detector."""
-        self.reachability_dim = reachability_dim
-        self.memory_size = memory_size
-        self.frontier_areas = set()
-        self.frontier_history = deque(maxlen=memory_size)
-        self.frontier_decay_time = 100  # Steps after which frontier boost decays
-        
-    def update_frontiers(self, newly_reachable: Set[Tuple[int, int]]):
-        """Update frontier areas with newly reachable positions."""
-        current_time = time.time()
-        
-        # Add new frontier areas
-        for pos in newly_reachable:
-            self.frontier_areas.add(pos)
-            self.frontier_history.append((pos, current_time))
-        
-        # Remove old frontier areas (decay over time)
-        current_frontiers = set()
-        cutoff_time = current_time - self.frontier_decay_time
-        
-        for pos, timestamp in self.frontier_history:
-            if timestamp > cutoff_time:
-                current_frontiers.add(pos)
-        
-        self.frontier_areas = current_frontiers
-    
-    def is_in_frontier(self, position: Tuple[float, float], tolerance: float = 24.0) -> bool:
-        """Check if position is in a frontier area."""
-        pos_x, pos_y = position
-        
-        for frontier_x, frontier_y in self.frontier_areas:
-            # Convert grid coordinates to pixel coordinates for comparison
-            frontier_pixel_x = frontier_x * 24.0
-            frontier_pixel_y = frontier_y * 24.0
-            
-            distance = math.sqrt((pos_x - frontier_pixel_x)**2 + (pos_y - frontier_pixel_y)**2)
-            if distance <= tolerance:
-                return True
-        
-        return False
-    
-    def get_frontier_count(self) -> int:
-        """Get current number of frontier areas."""
-        return len(self.frontier_areas)
-
-
-class StrategicWeighter:
-    """
-    Computes strategic weights for exploration based on proximity to level objectives.
-    
-    This component guides exploration toward level completion objectives (doors, switches, exit)
-    rather than purely random novelty seeking, improving sample efficiency on goal-directed tasks.
-    
-    References:
-    - Goal-directed exploration: Andrychowicz et al. (2017) "Hindsight Experience Replay"
-    - Strategic planning: Sutton & Barto (2018) "Reinforcement Learning: An Introduction"
-    """
-    
-    def __init__(self, reachability_dim: int = 64, objective_dim: int = 32):
-        """Initialize strategic weighter."""
-        self.reachability_dim = reachability_dim
-        self.objective_dim = objective_dim
-        
-        # Strategic weighting parameters
-        self.door_weight = 2.0      # Weight for proximity to doors
-        self.switch_weight = 1.5    # Weight for proximity to switches  
-        self.exit_weight = 3.0      # Weight for proximity to exit
-        self.distance_scale = 100.0 # Distance scaling factor (pixels)
-    
-    def compute_strategic_weight(self, 
-                               position: Tuple[float, float],
-                               door_positions: List[Tuple[float, float]] = None,
-                               switch_positions: List[Tuple[float, float]] = None,
-                               exit_position: Tuple[float, float] = None) -> float:
-        """
-        Compute strategic weight based on proximity to objectives.
-        
-        Args:
-            position: Current position (x, y)
-            door_positions: List of door positions
-            switch_positions: List of switch positions  
-            exit_position: Exit position
-            
-        Returns:
-            Strategic weight multiplier (>= 1.0)
-        """
-        if not any([door_positions, switch_positions, exit_position]):
-            return 1.0  # No strategic information available
-        
-        pos_x, pos_y = position
-        max_weight = 1.0
-        
-        # Weight by proximity to doors
-        if door_positions:
-            for door_x, door_y in door_positions:
-                distance = math.sqrt((pos_x - door_x)**2 + (pos_y - door_y)**2)
-                weight = self.door_weight * math.exp(-distance / self.distance_scale)
-                max_weight = max(max_weight, 1.0 + weight)
-        
-        # Weight by proximity to switches
-        if switch_positions:
-            for switch_x, switch_y in switch_positions:
-                distance = math.sqrt((pos_x - switch_x)**2 + (pos_y - switch_y)**2)
-                weight = self.switch_weight * math.exp(-distance / self.distance_scale)
-                max_weight = max(max_weight, 1.0 + weight)
-        
-        # Weight by proximity to exit
-        if exit_position:
-            exit_x, exit_y = exit_position
-            distance = math.sqrt((pos_x - exit_x)**2 + (pos_y - exit_y)**2)
-            weight = self.exit_weight * math.exp(-distance / self.distance_scale)
-            max_weight = max(max_weight, 1.0 + weight)
-        
-        return max_weight
-
-
-class ReachabilityPredictor(nn.Module):
-    """
-    Neural predictor for position reachability assessment within ICM framework.
-    
-    This component helps the ICM module make reachability assessments when detailed
-    reachability analysis from nclone is not available or too computationally expensive.
-    The predictor learns from ground truth reachability data during training.
-    
-    Architecture based on standard supervised learning for binary classification,
-    adapted for spatial reasoning tasks in NPP environments.
-    
-    References:
-    - Binary classification: Standard neural network literature
-    - Spatial reasoning: Hamilton et al. (2017) "Inductive Representation Learning on Large Graphs"
-    - Integration pattern: Custom design for ICM enhancement
-    """
-    
-    def __init__(self, observation_space, hidden_dim: int = 128):
-        """Initialize reachability predictor."""
-        super().__init__()
-        
-        # For simplicity, assume we get position coordinates as input
-        # In practice, this would extract relevant features from observation_space
-        self.input_dim = 4  # current_x, current_y, target_x, target_y
-        
-        # Prediction network
-        self.predictor = nn.Sequential(
-            nn.Linear(self.input_dim, hidden_dim),
-            nn.ReLU(),
-            nn.Dropout(0.1),
-            nn.Linear(hidden_dim, hidden_dim // 2),
-            nn.ReLU(),
-            nn.Dropout(0.1),
-            nn.Linear(hidden_dim // 2, 1),
-            nn.Sigmoid()  # Output probability of reachability
-        )
-    
-    def forward(self, current_pos: torch.Tensor, target_pos: torch.Tensor) -> torch.Tensor:
-        """
-        Predict reachability probability.
-        
-        Args:
-            current_pos: Current position [batch_size, 2]
-            target_pos: Target position [batch_size, 2]
-            
-        Returns:
-            Reachability probability [batch_size, 1]
-        """
-        # Concatenate positions
-        input_features = torch.cat([current_pos, target_pos], dim=1)
-        
-        # Predict reachability
-        reachability_prob = self.predictor(input_features)
-        
-        return reachability_prob
-    
-    def predict_reachability(self, current_pos: Tuple[float, float], target_pos: Tuple[float, float]) -> float:
-        """Predict reachability for a single position pair."""
-        with torch.no_grad():
-            current_tensor = torch.tensor([[current_pos[0], current_pos[1]]], dtype=torch.float32)
-            target_tensor = torch.tensor([[target_pos[0], target_pos[1]]], dtype=torch.float32)
-            
-            prob = self.forward(current_tensor, target_tensor)
-            return prob.item()
+from .reachability_exploration import ReachabilityAwareExplorationCalculator
 
 
 class ICMNetwork(nn.Module):
     """
-    Intrinsic Curiosity Module network with inverse and forward models and reachability integration.
-
-    The enhanced ICM consists of:
-    1. Inverse model: predicts action from state features φ(s_t) and φ(s_{t+1})
-    2. Forward model: predicts next state features φ(s_{t+1}) from φ(s_t) and action
-    3. Reachability modulation: scales curiosity based on accessibility analysis
-    4. Frontier detection: boosts exploration of newly accessible areas
-    5. Strategic weighting: prioritizes exploration near level objectives
+    Reachability-aware ICM that integrates with nclone's reachability systems.
     
-    Architecture Enhancements:
-    - Base ICM: Forward/inverse model prediction errors (existing functionality)
-    - Reachability Scaling: Modulate curiosity based on reachability analysis
-    - Frontier Detection: Boost curiosity for newly accessible areas
-    - Strategic Weighting: Prioritize exploration near level objectives
-    
-    Performance Optimizations:
-    - Lazy initialization of reachability components for dependency management
-    - Caching mechanisms for repeated reachability computations
-    - <1ms curiosity computation target for real-time RL training
-    
-    References:
-    - ICM Foundation: Pathak et al. (2017) "Curiosity-driven Exploration by Self-supervised Prediction"
-    - Reachability Guidance: Ecoffet et al. (2019) "Go-Explore: a New Approach for Hard-Exploration Problems"
-    - Strategic Exploration: Custom integration with nclone physics system
+    This implementation:
+    - Uses standard ICM forward/inverse models for base curiosity
+    - Integrates with nclone's reachability analysis for modulation
+    - Leverages existing ExplorationRewardCalculator logic
+    - Avoids duplicating exploration tracking functionality
     """
-
+    
     def __init__(
         self,
-        feature_dim: int,
-        action_dim: int,
+        feature_dim: int = 512,
+        action_dim: int = 6,
         hidden_dim: int = 256,
         eta: float = 0.01,
         lambda_inv: float = 0.1,
         lambda_fwd: float = 0.9,
         enable_reachability_awareness: bool = True,
-        reachability_dim: int = 64,
+        debug: bool = False
     ):
         """
-        Initialize enhanced ICM network with reachability awareness.
-
+        Initialize reachability-aware ICM.
+        
         Args:
-            feature_dim: Dimension of feature representations φ(s)
-            action_dim: Number of discrete actions (6 for N++)
-            hidden_dim: Hidden layer dimension for ICM networks
-            eta: Scaling factor for intrinsic reward
+            feature_dim: Dimension of feature representations
+            action_dim: Number of discrete actions
+            hidden_dim: Hidden layer dimension
+            eta: Intrinsic reward scaling factor
             lambda_inv: Weight for inverse model loss
             lambda_fwd: Weight for forward model loss
-            enable_reachability_awareness: Whether to enable reachability-aware modulation
-            reachability_dim: Dimension of reachability features (64 from nclone)
+            enable_reachability_awareness: Enable reachability modulation
+            debug: Enable debug output
         """
         super().__init__()
-
+        
         self.feature_dim = feature_dim
         self.action_dim = action_dim
+        self.hidden_dim = hidden_dim
         self.eta = eta
         self.lambda_inv = lambda_inv
         self.lambda_fwd = lambda_fwd
         self.enable_reachability_awareness = enable_reachability_awareness
-        self.reachability_dim = reachability_dim
-
-        # Inverse model: φ(s_t), φ(s_{t+1}) -> action distribution
+        self.debug = debug
+        
+        # Standard ICM components
         self.inverse_model = nn.Sequential(
             nn.Linear(feature_dim * 2, hidden_dim),
             nn.ReLU(),
             nn.Linear(hidden_dim, hidden_dim),
             nn.ReLU(),
-            nn.Linear(hidden_dim, action_dim),
+            nn.Linear(hidden_dim, action_dim)
         )
-
-        # Forward model: φ(s_t), action -> φ(s_{t+1})
+        
         self.forward_model = nn.Sequential(
             nn.Linear(feature_dim + action_dim, hidden_dim),
             nn.ReLU(),
             nn.Linear(hidden_dim, hidden_dim),
             nn.ReLU(),
-            nn.Linear(hidden_dim, feature_dim),
+            nn.Linear(hidden_dim, feature_dim)
         )
         
-        # Reachability-aware enhancement components (initialized if enabled)
-        if self.enable_reachability_awareness:
-            # Reachability scaling hyperparameters
-            # These parameters modulate base ICM curiosity based on accessibility
-            self.reachability_scale_factor = 2.0      # Boost for reachable areas
-            self.frontier_boost_factor = 3.0          # Extra boost for newly accessible areas
-            self.strategic_weight_factor = 1.5        # Weight for objective-proximate areas
-            self.unreachable_penalty = 0.1            # Penalty for confirmed unreachable areas
-            
-            # Initialize reachability components
-            self.frontier_detector = FrontierDetector(
-                reachability_dim=reachability_dim, memory_size=1000
-            )
-            self.strategic_weighter = StrategicWeighter(
-                reachability_dim=reachability_dim, objective_dim=32
-            )
-            self.exploration_history = ExplorationHistory(max_size=10000)
-            
-            # Track reachable positions for frontier detection
-            self.last_reachable_positions = set()
-
-    def forward(
-        self,
-        features_current: torch.Tensor,
-        features_next: torch.Tensor,
-        actions: torch.Tensor,
-        reachability_info: Optional[Dict[str, Any]] = None,
-    ) -> Dict[str, torch.Tensor]:
-        """
-        Forward pass through enhanced ICM networks with optional reachability awareness.
-
-        Args:
-            features_current: Current state features φ(s_t) [batch_size, feature_dim]
-            features_next: Next state features φ(s_{t+1}) [batch_size, feature_dim]
-            actions: Actions taken [batch_size] (discrete action indices)
-            reachability_info: Optional reachability information for modulation
-
-        Returns:
-            Dictionary containing:
-                - predicted_actions: Inverse model predictions [batch_size, action_dim]
-                - predicted_features: Forward model predictions [batch_size, feature_dim]
-                - intrinsic_reward: Intrinsic rewards [batch_size] (optionally reachability-modulated)
-                - inverse_loss: Inverse model loss
-                - forward_loss: Forward model loss
-                - reachability_modulation: Reachability modulation factors [batch_size] (if enabled)
-        """
-        # Inverse model: predict action from state features
+        # Reachability integration
+        if enable_reachability_awareness:
+            self.reachability_calculator = ReachabilityAwareExplorationCalculator(debug=debug)
+        else:
+            self.reachability_calculator = None
+        
+        # Performance tracking
+        self.computation_times = deque(maxlen=100)
+    
+    def forward(self, features_current: torch.Tensor, features_next: torch.Tensor, actions: torch.Tensor):
+        """Forward pass through ICM models."""
+        batch_size = features_current.shape[0]
+        
+        # Inverse model: predict action from state transition
         inverse_input = torch.cat([features_current, features_next], dim=1)
         predicted_actions = self.inverse_model(inverse_input)
-
-        # Forward model: predict next state features
-        actions_onehot = F.one_hot(actions.long(), num_classes=self.action_dim).float()
-        forward_input = torch.cat([features_current, actions_onehot], dim=1)
-        predicted_features = self.forward_model(forward_input)
-
-        # Compute losses
-        inverse_loss = F.cross_entropy(predicted_actions, actions.long())
-        forward_loss = F.mse_loss(predicted_features, features_next.detach())
-
-        # Compute base intrinsic reward as prediction error
-        prediction_error = torch.norm(
-            predicted_features - features_next.detach(), dim=1, p=2
-        )
-        base_intrinsic_reward = self.eta * 0.5 * prediction_error.pow(2)
         
-        # Apply reachability-aware modulation if enabled and information is available
-        reachability_modulation = torch.ones_like(base_intrinsic_reward)
+        # Forward model: predict next state from current state and action
+        actions_one_hot = F.one_hot(actions.long(), num_classes=self.action_dim).float()
+        forward_input = torch.cat([features_current, actions_one_hot], dim=1)
+        predicted_next_features = self.forward_model(forward_input)
         
-        if self.enable_reachability_awareness and reachability_info is not None:
-            reachability_modulation = self._compute_reachability_modulation(
-                reachability_info, base_intrinsic_reward.shape[0]
-            )
-        
-        # Final intrinsic reward with reachability modulation
-        intrinsic_reward = base_intrinsic_reward * reachability_modulation
-
-        result = {
-            "predicted_actions": predicted_actions,
-            "predicted_features": predicted_features,
-            "intrinsic_reward": intrinsic_reward,
-            "inverse_loss": inverse_loss,
-            "forward_loss": forward_loss,
-        }
-        
-        if self.enable_reachability_awareness:
-            result["reachability_modulation"] = reachability_modulation
-            
-        return result
-
-    def _compute_reachability_modulation(self, reachability_info: Dict[str, Any], batch_size: int) -> torch.Tensor:
-        """
-        Compute reachability-aware modulation factors for intrinsic rewards.
-        
-        This method implements the core reachability-aware curiosity enhancement,
-        combining reachability scaling, frontier boosting, and strategic weighting.
-        Optimized for <1ms computation time.
-        
-        Args:
-            reachability_info: Dictionary containing reachability information
-            batch_size: Size of the current batch
-            
-        Returns:
-            Modulation factors [batch_size] to multiply with base ICM rewards
-        """
-        device = next(self.parameters()).device
-        modulation_factors = torch.ones(batch_size, device=device)
-        
-        # Extract reachability information with early exit for performance
-        current_positions = reachability_info.get('current_positions', [])
-        target_positions = reachability_info.get('target_positions', [])
-        reachable_positions = reachability_info.get('reachable_positions', [])
-        
-        # Early exit if no position information
-        if not current_positions or not target_positions:
-            return modulation_factors
-        
-        # Vectorized computation for better performance
-        valid_indices = min(len(current_positions), len(target_positions), batch_size)
-        
-        for i in range(valid_indices):
-            target_pos = target_positions[i]
-            reachable_set = set(reachable_positions[i]) if i < len(reachable_positions) else set()
-            
-            # Fast reachability check (simplified for performance)
-            target_grid_x = int(target_pos[0] / 24.0)
-            target_grid_y = int(target_pos[1] / 24.0)
-            target_grid_pos = (target_grid_x, target_grid_y)
-            
-            # Simple reachability scaling
-            if target_grid_pos in reachable_set:
-                modulation_factors[i] = self.reachability_scale_factor
-            else:
-                # Check if on frontier (simplified)
-                is_frontier = any(
-                    (target_grid_x + dx, target_grid_y + dy) in reachable_set
-                    for dx in [-1, 0, 1] for dy in [-1, 0, 1]
-                    if dx != 0 or dy != 0
-                )
-                modulation_factors[i] = 0.5 if is_frontier else self.unreachable_penalty
-        
-        return modulation_factors
+        return predicted_actions, predicted_next_features
     
-    def _compute_reachability_scale(self, current_pos: Tuple[float, float], 
-                                  target_pos: Tuple[float, float], 
-                                  reachable_positions: Set[Tuple[int, int]]) -> float:
-        """
-        Scale curiosity based on reachability status.
-        
-        Scaling Strategy:
-        - Reachable areas: 1.0x (full curiosity)
-        - Frontier areas: 0.5x (moderate curiosity)  
-        - Unreachable areas: 0.1x (minimal curiosity)
-        """
-        # Convert target position to grid coordinates for reachability check
-        target_grid_x = int(target_pos[0] / 24.0)  # 24 pixel cells
-        target_grid_y = int(target_pos[1] / 24.0)
-        target_grid_pos = (target_grid_x, target_grid_y)
-        
-        # Check if target position is reachable
-        if target_grid_pos in reachable_positions:
-            return 1.0  # Full curiosity for reachable areas
-        
-        # Check if target is on exploration frontier
-        if self._is_on_exploration_frontier(target_grid_pos, reachable_positions):
-            return 0.5  # Moderate curiosity for frontier
-        
-        # Unreachable area gets penalty
-        return self.unreachable_penalty
-    
-    def _compute_frontier_boost(self, target_pos: Tuple[float, float], 
-                              current_reachable: Set[Tuple[int, int]]) -> float:
-        """
-        Boost curiosity for newly reachable areas (frontiers).
-        """
-        # Detect newly reachable areas
-        newly_reachable = current_reachable - self.last_reachable_positions
-        
-        if newly_reachable:
-            # Update frontier detector
-            self.frontier_detector.update_frontiers(newly_reachable)
-            
-            # Check if current exploration is in frontier area
-            if self.frontier_detector.is_in_frontier(target_pos):
-                return self.frontier_boost_factor
-        
-        # Update last reachable positions
-        self.last_reachable_positions = current_reachable.copy()
-        
-        return 1.0  # No boost
-    
-    def _is_on_exploration_frontier(self, target_pos: Tuple[int, int], 
-                                  reachable_positions: Set[Tuple[int, int]]) -> bool:
-        """Check if position is on the exploration frontier (adjacent to reachable areas)."""
-        target_x, target_y = target_pos
-        
-        # Check 8-connected neighbors
-        for dx in [-1, 0, 1]:
-            for dy in [-1, 0, 1]:
-                if dx == 0 and dy == 0:
-                    continue
-                neighbor = (target_x + dx, target_y + dy)
-                if neighbor in reachable_positions:
-                    return True
-        
-        return False
-
     def compute_intrinsic_reward(
         self,
         features_current: torch.Tensor,
         features_next: torch.Tensor,
         actions: torch.Tensor,
-        reachability_info: Optional[Dict[str, Any]] = None,
+        observations: Optional[Dict[str, Any]] = None
     ) -> torch.Tensor:
         """
-        Compute intrinsic reward without computing gradients, with optional reachability awareness.
-
+        Compute intrinsic reward with optional reachability modulation.
+        
         Args:
             features_current: Current state features
-            features_next: Next state features
+            features_next: Next state features  
             actions: Actions taken
-            reachability_info: Optional reachability information for modulation
-
-        Returns:
-            Intrinsic rewards [batch_size] (optionally reachability-modulated)
-        """
-        with torch.no_grad():
-            actions_onehot = F.one_hot(
-                actions.long(), num_classes=self.action_dim
-            ).float()
-            forward_input = torch.cat([features_current, actions_onehot], dim=1)
-            predicted_features = self.forward_model(forward_input)
-
-            prediction_error = torch.norm(
-                predicted_features - features_next, dim=1, p=2
-            )
-            base_intrinsic_reward = self.eta * 0.5 * prediction_error.pow(2)
+            observations: Environment observations for reachability analysis
             
-            # Apply reachability-aware modulation if enabled and information is available
-            if self.enable_reachability_awareness and reachability_info is not None:
-                reachability_modulation = self._compute_reachability_modulation(
-                    reachability_info, base_intrinsic_reward.shape[0]
+        Returns:
+            Intrinsic reward tensor
+        """
+        start_time = time.time()
+        
+        # Compute base ICM curiosity (forward model prediction error)
+        with torch.no_grad():
+            _, predicted_next_features = self.forward(features_current, features_next, actions)
+            prediction_error = F.mse_loss(predicted_next_features, features_next, reduction='none')
+            base_curiosity = prediction_error.mean(dim=1) * self.eta
+        
+        # Apply reachability modulation if enabled and available
+        # Skip if no level data to avoid expensive failed computations
+        if (self.enable_reachability_awareness and 
+            self.reachability_calculator is not None and 
+            observations is not None and
+            observations.get("level_data") is not None):
+            
+            try:
+                modulated_curiosity = self._apply_reachability_modulation(
+                    base_curiosity, observations
                 )
-                intrinsic_reward = base_intrinsic_reward * reachability_modulation
+            except Exception as e:
+                if self.debug:
+                    print(f"Reachability modulation failed: {e}")
+                modulated_curiosity = base_curiosity
+        else:
+            modulated_curiosity = base_curiosity
+        
+        # Track performance
+        computation_time = (time.time() - start_time) * 1000
+        self.computation_times.append(computation_time)
+        
+        return modulated_curiosity
+    
+    def _apply_reachability_modulation(
+        self,
+        base_curiosity: torch.Tensor,
+        observations: Dict[str, Any]
+    ) -> torch.Tensor:
+        """Apply reachability-aware modulation using nclone systems."""
+        batch_size = base_curiosity.shape[0]
+        modulated_curiosity = base_curiosity.clone()
+        
+        for i in range(batch_size):
+            # Extract single observation
+            obs_i = self._extract_single_observation(observations, i)
+            
+            # Get reachability analysis from nclone
+            reward_info = self.reachability_calculator.calculate_reachability_aware_reward(
+                player_x=obs_i.get("player_x", 0.0),
+                player_y=obs_i.get("player_y", 0.0),
+                level_data=obs_i.get("level_data"),
+                switch_states=obs_i.get("switch_states", {})
+            )
+            
+            # Apply modulation factor
+            modulation_factor = reward_info.get("reachability_modulation", 1.0)
+            modulated_curiosity[i] *= modulation_factor
+        
+        return modulated_curiosity
+    
+    def _extract_single_observation(self, observations: Dict[str, Any], index: int) -> Dict[str, Any]:
+        """Extract single observation from batch."""
+        single_obs = {}
+        for key, value in observations.items():
+            if isinstance(value, (list, np.ndarray, torch.Tensor)):
+                if len(value) > index:
+                    single_obs[key] = value[index]
+                else:
+                    single_obs[key] = value[0] if len(value) > 0 else None
             else:
-                intrinsic_reward = base_intrinsic_reward
-
-        return intrinsic_reward
+                single_obs[key] = value
+        return single_obs
+    
+    def compute_losses(
+        self,
+        features_current: torch.Tensor,
+        features_next: torch.Tensor,
+        actions: torch.Tensor
+    ) -> Dict[str, torch.Tensor]:
+        """Compute ICM losses."""
+        predicted_actions, predicted_next_features = self.forward(
+            features_current, features_next, actions
+        )
+        
+        # Inverse model loss
+        inverse_loss = F.cross_entropy(predicted_actions, actions.long())
+        
+        # Forward model loss
+        forward_loss = F.mse_loss(predicted_next_features, features_next)
+        
+        # Total loss
+        total_loss = self.lambda_inv * inverse_loss + self.lambda_fwd * forward_loss
+        
+        return {
+            "total_loss": total_loss,
+            "inverse_loss": inverse_loss,
+            "forward_loss": forward_loss
+        }
+    
+    def get_reachability_info(self, observations: Dict[str, Any]) -> Dict[str, Any]:
+        """Get reachability information using nclone systems."""
+        if self.reachability_calculator is None:
+            return {"available": False}
+        
+        try:
+            # Extract first observation
+            obs = self._extract_single_observation(observations, 0)
+            
+            # Get compact features
+            compact_features = self.reachability_calculator.extract_compact_features(
+                level_data=obs.get("level_data"),
+                player_position=(obs.get("player_x", 0.0), obs.get("player_y", 0.0)),
+                switch_states=obs.get("switch_states", {})
+            )
+            
+            # Get frontier information
+            frontiers = self.reachability_calculator.get_frontier_information(
+                level_data=obs.get("level_data"),
+                player_position=(obs.get("player_x", 0.0), obs.get("player_y", 0.0)),
+                switch_states=obs.get("switch_states", {})
+            )
+            
+            return {
+                "available": True,
+                "compact_features": compact_features,
+                "frontiers": frontiers,
+                "num_frontiers": len(frontiers)
+            }
+            
+        except Exception as e:
+            if self.debug:
+                print(f"Reachability info extraction failed: {e}")
+            return {"available": False, "error": str(e)}
+    
+    def reset(self):
+        """Reset for new episode."""
+        if self.reachability_calculator is not None:
+            self.reachability_calculator.reset()
+        self.computation_times.clear()
+    
+    def get_performance_stats(self) -> Dict[str, float]:
+        """Get performance statistics."""
+        if len(self.computation_times) == 0:
+            return {"avg_time_ms": 0.0, "max_time_ms": 0.0, "min_time_ms": 0.0}
+        
+        times = np.array(self.computation_times)
+        return {
+            "avg_time_ms": float(np.mean(times)),
+            "max_time_ms": float(np.max(times)),
+            "min_time_ms": float(np.min(times)),
+            "p95_time_ms": float(np.percentile(times, 95)),
+            "samples": len(times)
+        }
 
 
 class ICMTrainer:
     """
-    Trainer for ICM that handles optimization and logging.
+    Trainer for reachability-aware ICM.
+    
+    This trainer integrates with nclone's exploration systems and provides
+    a clean interface for RL training loops.
     """
-
+    
     def __init__(
-        self, icm_network: ICMNetwork, learning_rate: float = 1e-3, device: str = "cpu"
+        self,
+        icm_network: ICMNetwork,
+        learning_rate: float = 1e-3,
+        device: str = "cpu"
     ):
         """
         Initialize ICM trainer.
-
+        
         Args:
             icm_network: ICM network to train
-            learning_rate: Learning rate for ICM optimizer
+            learning_rate: Learning rate for optimizer
             device: Device to run on
         """
         self.icm_network = icm_network.to(device)
-        self.optimizer = torch.optim.Adam(
-            self.icm_network.parameters(), lr=learning_rate
-        )
+        self.optimizer = torch.optim.Adam(icm_network.parameters(), lr=learning_rate)
         self.device = device
-
-        # Logging
-        self.train_stats = {
-            "inverse_loss": [],
-            "forward_loss": [],
-            "total_loss": [],
-            "mean_intrinsic_reward": [],
-        }
-
+    
     def update(
         self,
         features_current: torch.Tensor,
         features_next: torch.Tensor,
         actions: torch.Tensor,
-        reachability_info: Optional[Dict[str, Any]] = None,
+        observations: Optional[Dict[str, Any]] = None
     ) -> Dict[str, float]:
         """
-        Update ICM networks with a batch of experience, with optional reachability awareness.
-
+        Update ICM with a batch of transitions.
+        
         Args:
             features_current: Current state features
             features_next: Next state features
             actions: Actions taken
-            reachability_info: Optional reachability information for modulation
-
+            observations: Environment observations
+            
         Returns:
-            Dictionary of training statistics
+            Training statistics
         """
         # Move to device
         features_current = features_current.to(self.device)
         features_next = features_next.to(self.device)
         actions = actions.to(self.device)
-
-        # Forward pass with optional reachability information
-        icm_output = self.icm_network(features_current, features_next, actions, reachability_info)
-
-        # Compute total loss
-        total_loss = (
-            self.icm_network.lambda_inv * icm_output["inverse_loss"]
-            + self.icm_network.lambda_fwd * icm_output["forward_loss"]
-        )
-
-        # Backward pass
+        
+        # Compute losses
+        losses = self.icm_network.compute_losses(features_current, features_next, actions)
+        
+        # Update parameters
         self.optimizer.zero_grad()
-        total_loss.backward()
+        losses["total_loss"].backward()
         self.optimizer.step()
-
-        # Log statistics
+        
+        # Return statistics
         stats = {
-            "inverse_loss": icm_output["inverse_loss"].item(),
-            "forward_loss": icm_output["forward_loss"].item(),
-            "total_loss": total_loss.item(),
-            "mean_intrinsic_reward": icm_output["intrinsic_reward"].mean().item(),
+            "total_loss": losses["total_loss"].item(),
+            "inverse_loss": losses["inverse_loss"].item(),
+            "forward_loss": losses["forward_loss"].item()
         }
-
-        for key, value in stats.items():
-            self.train_stats[key].append(value)
-
+        
+        # Add performance stats
+        perf_stats = self.icm_network.get_performance_stats()
+        stats.update(perf_stats)
+        
         return stats
-
+    
     def get_intrinsic_reward(
         self,
         features_current: torch.Tensor,
         features_next: torch.Tensor,
         actions: torch.Tensor,
-        reachability_info: Optional[Dict[str, Any]] = None,
-    ) -> np.ndarray:
-        """
-        Get intrinsic rewards for a batch of transitions, with optional reachability awareness.
-
-        Args:
-            features_current: Current state features
-            features_next: Next state features
-            actions: Actions taken
-            reachability_info: Optional reachability information for modulation
-
-        Returns:
-            Intrinsic rewards as numpy array (optionally reachability-modulated)
-        """
+        observations: Optional[Dict[str, Any]] = None
+    ) -> torch.Tensor:
+        """Get intrinsic rewards for a batch of transitions."""
         features_current = features_current.to(self.device)
         features_next = features_next.to(self.device)
         actions = actions.to(self.device)
-
-        intrinsic_reward = self.icm_network.compute_intrinsic_reward(
-            features_current, features_next, actions, reachability_info
+        
+        return self.icm_network.compute_intrinsic_reward(
+            features_current, features_next, actions, observations
         )
+    
+    def reset(self):
+        """Reset for new episode."""
+        self.icm_network.reset()
 
-        return intrinsic_reward.cpu().numpy()
 
-    def get_recent_stats(self, window: int = 100) -> Dict[str, float]:
-        """
-        Get recent training statistics.
-
-        Args:
-            window: Number of recent updates to average over
-
-        Returns:
-            Dictionary of averaged statistics
-        """
-        stats = {}
-        for key, values in self.train_stats.items():
-            if len(values) > 0:
-                recent_values = values[-window:]
-                stats[key] = np.mean(recent_values)
-            else:
-                stats[key] = 0.0
-
-        return stats
+# Backward compatibility aliases
+ConsolidatedICMNetwork = ICMNetwork
+ConsolidatedICMTrainer = ICMTrainer
